+++ Source/SwiftLintBuiltInRules/Rules/RuleConfigurations/MultilineParametersConfiguration.swift
@@ -9,8 +9,21 @@ struct MultilineParametersConfiguration: SeverityBasedRuleConfiguration {
     @ConfigurationElement(key: "max_number_of_single_line_parameters")
     private(set) var maxNumberOfSingleLineParameters: Int?
 
+    @ConfigurationElement(key: "requires_single_line")
+    private(set) var requiresSingleLine = false
+
     // swiftlint:disable:next unneeded_throws_rethrows
     func validate() throws(Issue) {
+        if requiresSingleLine, !allowsSingleLine {
+            Issue.inconsistentConfiguration(
+                ruleID: Parent.identifier,
+                message: """
+                         Option '\($requiresSingleLine.key)' cannot be true when \
+                         '\($allowsSingleLine.key)' is false.
+                         """
+            ).print()
+        }
+
         guard let maxNumberOfSingleLineParameters else {
             return
         }